requirements.txt
customtkinter
pycryptodome
cryptography
